requirements.txt
anyio==4.4.0
astroid==3.2.4
certifi==2024.8.30
charset-normalizer==3.3.2
dill==0.3.8
idna==3.10
iqoptionapi @ https://github.com/iqoptionapi/iqoptionapi/archive/refs/heads/master.zip#sha256=25326cba8e21bc3bb082ab2bd21282f435ce25589d89499a92e1879559c4ca42
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
packaging==24.1
pip-review==1.3.0
platformdirs==4.3.6
Pygments==2.18.0
pylint==3.2.7
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
six==1.16.0
sniffio==1.3.1
tinydb==4.8.0
tomlkit==0.13.2
urllib3==2.2.3
websocket-client==0.56.0
websockets==13.0.1
